ahora al añadir un order se puede elegir la cantidad prepagada

+++ ASCIITeam-ERP-2018-2019/ERP-ASCIITeam/ERP/Dominio/Order.cs
@@ -26,7 +26,7 @@ namespace ERP.Dominio
             this.gestorOrder = new GestorOrder();
         }
 
-        public Order(int iD, decimal refCustomer, decimal refUser, DateTime datet, int refPaymentMethod, decimal total, int prepaid, int deleted)
+        public Order(int iD, decimal refCustomer, decimal refUser, DateTime datet, int refPaymentMethod, decimal total, decimal prepaid, int deleted)
         {
             this.id = iD;
             this.refCustomer = refCustomer;
